requirements.txt
selenium==4.27.1
numpy==2.2.1 
matplotlib
pandas==2.2.3
scipy==1.14.1
seaborn==0.13.2
ipykernel==6.29.5
jupyter==1.1.1
notebook==7.3.2
python-dotenv==1.2.2
